+++ lib/src/list.dart
@@ -1,31 +1,62 @@
-import 'package:flutter/widgets.dart';
+part of '../styled_widget.dart';
 
 extension StyledList<E> on List<Widget> {
   Widget toColumn({
+    Key key,
     MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+    MainAxisSize mainAxisSize = MainAxisSize.max,
     CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+    TextDirection textDirection,
+    VerticalDirection verticalDirection = VerticalDirection.down,
+    TextBaseline textBaseline,
+    List<Widget> children = const <Widget>[],
   }) =>
       Column(
+        key: key,
         mainAxisAlignment: mainAxisAlignment,
+        mainAxisSize: mainAxisSize,
         crossAxisAlignment: crossAxisAlignment,
+        textDirection: textDirection,
+        verticalDirection: verticalDirection,
+        textBaseline: textBaseline,
         children: this,
       );
 
   Widget toRow({
+    Key key,
     MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+    MainAxisSize mainAxisSize = MainAxisSize.max,
     CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+    TextDirection textDirection,
+    VerticalDirection verticalDirection = VerticalDirection.down,
+    TextBaseline textBaseline,
+    List<Widget> children = const <Widget>[],
   }) =>
       Row(
+        key: key,
         mainAxisAlignment: mainAxisAlignment,
+        mainAxisSize: mainAxisSize,
         crossAxisAlignment: crossAxisAlignment,
+        textDirection: textDirection,
+        verticalDirection: verticalDirection,
+        textBaseline: textBaseline,
         children: this,
       );
 
   Widget toStack({
+    Key key,
     AlignmentGeometry alignment = AlignmentDirectional.topStart,
+    TextDirection textDirection,
+    StackFit fit = StackFit.loose,
+    Overflow overflow = Overflow.clip,
+    List<Widget> children = const <Widget>[],
   }) =>
       Stack(
+        key: key,
         alignment: alignment,
+        textDirection: textDirection,
+        fit: fit,
+        overflow: overflow,
         children: this,
       );
 }