+++ lib/src/text_span.dart
@@ -1,5 +1,4 @@
-import 'package:flutter/gestures.dart';
-import 'package:flutter/widgets.dart';
+part of '../styled_widget.dart';
 
 extension StyledTextSpan on TextSpan {
   TextSpan copyWith({