+++ lib/styled_widget.dart
@@ -1,12 +1,18 @@
 library styled_widget;
 
-/*
-A styled widget should have its own respective file
+import 'dart:io';
+import 'dart:math';
+import 'dart:ui';
 
-*/
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/foundation.dart';
 
-export 'src/widget.dart';
-export 'src/text.dart';
-export 'src/text_span.dart';
-export 'src/icon.dart';
-export 'src/list.dart';
+part 'src/widget.dart';
+part 'src/text.dart';
+part 'src/text_span.dart';
+part 'src/list.dart';
+part 'src/icon.dart';
+part 'src/animated_widget.dart';
+part 'src/animated_text.dart';